Cleaning up. Add support for absolute mode for setting/getting feature values.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1983 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUIIDC++/testcam/MyCmdWindow.h
@@ -191,7 +191,7 @@ MyCmdWindow<CAMERA, PIXEL>::initializeMovie()
 
     _canvas.resize();
 
-    int	props[] = {0, _movie.nframes() - 1, 1};
+    int	props[] = {0, static_cast<int>(_movie.nframes() - 1), 1};
     _captureCmd.setProp(c_StatusMovie, props);
     
     repaintCanvas();
